feature: disable signup and account deletion in production

+++ lore-backend/src/users/users.controller.ts
@@ -21,6 +21,7 @@ import { UserByUsernamePipe } from "../common/pipes/user-by-username.pipe";
 import { AuthGuard } from "../auth/auth.guard";
 import { AuthService } from "../auth/auth.service";
 import { Response } from "express";
+import * as process from "node:process";
 
 @Controller("users")
 export class UsersController {
@@ -32,6 +33,9 @@ export class UsersController {
 
     @Post()
     create(@Body() createUserDto: CreateUserDto) {
+        if (process.env.NODE_ENV !== "development") {
+            throw new HttpException("", HttpStatus.FORBIDDEN);
+        }
         return this.usersService.create(createUserDto);
     }
 
@@ -47,6 +51,9 @@ export class UsersController {
     @UseGuards(AuthGuard)
     @Delete(":username")
     async remove(@Payload(UserByTokenPipe) user: User, @Res({ passthrough: true }) response: Response) {
+        if (process.env.NODE_ENV !== "development") {
+            throw new HttpException("", HttpStatus.FORBIDDEN);
+        }
         await this.usersService.remove(user);
         this.authService.logOut(response);
     }
